requirements.txt
Django==4.2
Pillow==9.5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-environ==0.10.0
djangorestframework-simplejwt==5.2.2
Faker==18.7.0
factory-boy==3.2.1
Pillow==9.5.0
django-filter==23.2